pyproject.toml
[tool.poetry]
name = "uniprot-viewer"
version = "0.1.0"
description = "A simple Django rest framework app to provide API for uniprot dataset"
authors = ["Scott Greig <scottjgreig@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
